requirements.txt
pydantic
fastapi
google-genai
cohere
python-dotenv
sqlalchemy
psycopg2-binary
alembic
uvicorn
passlib[bcrypt]
python-jose[cryptography]